#850 add git backend handler
The git-http-backend [guide](https://git-scm.com/docs/git-http-backend)  describes to match the request url against the "git-(upload|receive)-pack$". Unfortunately this is part of the url encoded query parameter, which as it seams Express does not use to match the regex.

Example request url:
https://alice.localhost:8443/public/test.git/info/refs?service=git-upload-pack

What Express uses to match the regex:
https://alice.localhost:8443/public/test.git/info/refs

Therefore all GET and POST request matching "/*.git*" are routed to the git backend. Where the request url are matched against git-upload-pack and git-receive-pack. Further, if the request does not match any of the two services, it is forwarded with the exact same route chain defended in the middleware handler. This allows us to serve the git protocol and the files, such that they are consumable by solid data browser.